% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.functions.r
\name{dublogistic.f}
\alias{dublogistic.f}
\title{A double logistic selectivity function}
\usage{
dublogistic.f(L, inflection1 = 38, inflection2 = 90, slope1 = 0.3,
  slope2 = 0.1, max.sel = 1, minsel.upper = 0.4, plot = F)
}
\arguments{
\item{L}{a length or length vector for which you want a selectivity estimates}

\item{inflection1}{the inflection point of the first part of the selectivity curve}

\item{inflection2}{the inflection point of the second part of the selectivity curve}

\item{slope1}{the slope of the first part of the curve}

\item{slope2}{the slope of the second part of the curve}

\item{max.sel}{the maximum selectivity level}

\item{minsel.upper}{the asymptotic minimum selectivity on the right side of the curve. Dome shaped curves between 0 and max.sel, =max.sel for a regular logistic curve}

\item{plot}{T if you want a plot of the curve}
}
\description{
This is a useful function because it can not only mimic standard sigmoidal selectivity but also
        a dome shaped selectivity curve if necessary. Set minsell.upper=max.sel for a simple logisitic curve. 
        A default parameterisation is provided simply to give you something to start with that might seem half
        decent.
}
